# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API's Room 911",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://servicios.asstiseguridadsocial.com"
        },
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/v1/employees?id={id}&department={department}": {
            "get": {
                "tags": [
                    "Employee"
                ],
                "summary": "Listar empleados",
                "operationId": "500dbc5a58b7146672787740b2a1a421",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "department",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "_rel": "employees",
                                        "_embedded": []
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Mensaje de error"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/employees": {
            "post": {
                "tags": [
                    "Employee"
                ],
                "summary": "Crear empleado",
                "operationId": "3d18b48472b00799cf4d9017a6a8938f",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "identification",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Empleado creado con éxito"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Mensaje de error"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/employees/{id}": {
            "put": {
                "tags": [
                    "Employee"
                ],
                "summary": "Editar empleado",
                "operationId": "559b2b563d089c94dceeb64d71d14293",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "department_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Empleado actualizado con éxito"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Mensaje de error"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Employee"
                ],
                "summary": "Eliminar un empleados",
                "operationId": "84e98370e7146e8736d76dcc536bed0a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Empleado eliminado satisfactoriamente"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Mensaje de error"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Employee"
                ],
                "summary": "Cambio de estado de un empleados",
                "operationId": "b6545520b8f7f75b8ef4b3217c558e28",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "state",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Cambio de estado realizado correctamente"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Mensaje de error"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/employees/bulk_upload_users": {
            "post": {
                "tags": [
                    "Employee"
                ],
                "summary": "Carga masiva de empleados",
                "operationId": "3ffd7a4038572a8eb70f65de387161f0",
                "parameters": [
                    {
                        "name": "file",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "file"
                        }
                    },
                    {
                        "name": "department_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Empleado actualizado con éxito"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Mensaje de error"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}